pyproject.toml
[tool.poetry]
name = "vulnbox"
version = "0.1.0"
description = ""
authors = ["Markus Bauer <markus.bauer@cispa.saarland>"]
packages = [
    { include = "vulnbuild" }
]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pyyaml = "^6.0.1"
python-hcl2 = "^4.3.3"
doit = "^0.36.0"

[tool.poetry.dev-dependencies]
mypy = "^1.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
types-requests = "^2.31.0.20240311"
types-pyyaml = "^6.0.12.20240311"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
files = ["vulnbuild", "tests", "resources/router-setup-scripts/*.py"]
python_version = "3.10"
disallow_untyped_defs = true

# === Python code ===
[tool.poetry.scripts]
vulnbuild = "vulnbuild.__main__:main"
